requirement.txt
pandas
numpy
scikit-learn
mlflow
tensorflow
pyyaml
dvc
dagshub